image-scraper/requirements.txt
requests==2.31.0  # HTTP library for web requests
beautifulsoup4==4.12.2  # HTML parsing for web scraping
lxml==4.9.3  # XML/HTML parser for BeautifulSoup
PyYAML==6.0.1  # YAML configuration file parsing
pytest==7.4.3  # Testing framework
pytest-cov==4.1.0  # Test coverage reporting
